// static/radar/src/types.d.ts
export interface components {
    schemas: {
        Bot: {
            identifier: string;
            name: string;
            /** @enum {string} */
            status: "IDLE" | "MOVING" | "MINING";
            coordinates: components["schemas"]["Coordinates"];
            inventory: number;
        };
        Coordinates: {
            /** Format: double */
            x: number;
            /** Format: double */
            y: number;
        };
        Error: string;
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}